Serialize ids and pairs as their names
Every id-keyed and pair-keyed container carried a hand-written serializer that
converted to a string-keyed std::map on save and rebuilt the ids on load.
Giving NameId and OrderedIdPair their own cereal and YAML converters makes the
conversion happen once, at the element: seven save/load pairs collapse back to
upstream's serialize() one-liners and the wire format is unchanged.

A pair writes its names alphabetically rather than in its canonical order,
which is decided by hash value and would otherwise tie every archive, document
and saved SRDF to the hashing implementation. orderedNameView() is the
accessor for that, and every site that renders a pair for output now uses it.

NameIdValue becomes std::size_t hashed by boost::hash, which is consistent
across platforms and avalanching, so boost's open-addressing containers skip
their post-mix. Hashing moves out of line to keep types.h free of boost.

+++ common/include/tesseract/common/fwd.h
@@ -25,12 +25,13 @@
 #ifndef TESSERACT_COMMON_TESSERACT_COMMON_FWD_H
 #define TESSERACT_COMMON_TESSERACT_COMMON_FWD_H
 
+#include <cstddef>
 #include <cstdint>
 
 namespace tesseract::common
 {
 // types.h — identity types (NameId family)
-using NameIdValue = std::uint64_t;
+using NameIdValue = std::size_t;
 struct LinkTag;
 struct JointTag;
 template <typename Tag>
